Active page tab - Change background property for IE11
Using a raw svg in the URL parameter of a background property in IE is a tricky thing. For example, background: url('data:image/svg+xml;utf8,<svg ...> ... </svg>'); will work in popular browsers but IE won't tolerate it. For it to work, we have to HTML encode any URL-unsavfe characters like <, > or ?, swap attribute values' double quotes with single quotes, and wrap the URL with double quotes. For more, see https://codepen.io/tigt/post/optimizing-svgs-in-data-uris