CSS - Change default values, remove deprecated property
Overflow-wrap was originally a nonstandard and unprefixed Microsoft extension called word-wrap, and was implemented by most browsers with the same name. It has since been renamed to overflow-wrap, with word-wrap being an alias. Therefore, we only need one and we should stick to overflow-wrap. This commit also changes the default values for hyphenation properties.