Update catch-up.md
Updated, based on the 17th meeting where I was re-discovering how the process works. I felt the instructions could be clearer about what to do with the PR's when reviewing.  The change to the second statement around when to look at backlog or not felt like a way to stop us reviewing both the to do column AND the backlog at every single meeting?